popup-notification: decrease font-size
When trying to install uBlock origin as suggested on the start page, I
found that the text of the popup is taking up so much space that the
"Add" / "Cancel" buttons are missing again. Decrease the font size even
further (like in the "Site information"), so the buttons are visible
again.

If somebody wants to improve this: maybe we could give the
popup-notification-body a fixed height instead, and make it scrollable?
Patches welcome, and see README.md for instructions how to use the
Firefox developer tools with the browser UI itself.