homepage: update note on how to install add-ons
addons.mozilla.org displays the following message when visiting it with
the Android Firefox user agent now, while also graying out the usual
button to install the add-on:
	You can install this add-on in the Add-ons Manager. Learn more
	about add-ons for Android.
It is not possible to use a different user agent just for this website
anymore, as it used to be with "general.useragent.site_specific_overrides"
(see https://bugzilla.mozilla.org/show_bug.cgi?id=1589607). Therefore,
just explain that one can press the "Download file" button to install
add-ons.