contrib/extract-search-engines.sh: new file (MR 19)
Add a script that allows extracting the list of search engines enabled
in the current firefox build. There is no way to simply disable all
built-in search engines, so we need a list like this. Note that the
search engines can only be removed in the ESR version. I've briefly
tried to find such a list in the source code, but didn't find it without
spending too much effort. Besides, I'm guessing that in the source code
it's stored in almost the same format found in the "omni.ja" zip
archive.