userChrome/tabmenu: make it work with private mode (MR 19)
Always reserver the space for the private browsing indicator, even if
not in private mode. This allows using the same width for the displayed
tab in both modes. Without this, the X of the tab is not visible in
private mode.

I think it's nice to have this space in both modes to quickly close the
menu, as it's right above the menu button.